Always find/insert nodes exactly at splice boundaries
Instead of recycling nodes that start or end existing changes, we just
always split existing changes by inserting a new node and updating
change text appropriately. This is clearer and will make it easy to
suppress combination of changes, which is needed for some applications.
It does have the downside of requiring us to create and delete extra
nodes when performing overlapping splices, but I think it’s worth it.