Increase randomness in the inputs
Previous tests were not catching the bad behavior we were observing in
text-buffer. The reason was that we were never splicing an empty new
extent in the reference patch, and we were missing out on some
interesting edge cases.

Also, since this data structure is so important, we've increased the
number of iterations we run on CI, so that we're 100% sure to have
validated all the inputs.