Avoid temporarily moving from text in text_diff
This move was done to avoid a copy. But text_diff was called in a
background thread, it was possible for other threads to observe
the temporarily mutated text. Now, we introduce a constructor for
VectorStringAdapter that borrows its content so that we can avoid
moving from the Text while also avoiding a copy.