Make most TextBuffer APIs deal w/ strings instead of Text
The line offsets array was not being used in most cases, so it was unnecessary to
include it. This extracts a type alias Text::String for Text's content vector.
TextBuffer methods like text_in_range and line_for_row now return these objects
so that we don't have to construct an unnecessary line_offsets array