Add the ability to save buffers to file descriptor integers
This adds type detection on the argument to TextBuffer.prototype.save.
If called with an integer, we assume it's a file descriptor.

I still want to verify that we can write to the given file descriptor,
since we close the file handle when the SaveWorker terminates and I'm
not clear how that relates to the file descriptor.